requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.10.3
requests==2.31.0
openai==1.3.6
replicate==0.22.0
supabase==2.10.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.0
rembg==2.0.57
Pillow==10.1.0
google-generativeai==0.8.3
scipy==1.14.1
numpy==1.26.4 